refactor(api): split routes into separate modules
- Create app/routes/ package with dedicated route files
- Move topics routes to app/routes/topics.py
- Move audiences routes to app/routes/audiences.py
- Move audience_templates routes to app/routes/audience_templates.py
- Simplify main.py to only FastAPI setup and router includes
